Split up implementation and instantiation
